requirements.txt
requests
dnspython
prettytable
